finished find_moves with signifficant speed gain! started make_moves

+++ src/chess_implementationC/chess_board.h
@@ -23,13 +23,13 @@ struct ChessBoard
     bool white_piece_alive[30];
     signed char white_piece_jump_moves[30][30];
     signed char white_piece_move_directions[30][28];
-    int white_piece_fist_move[30];
+    short white_piece_fist_move[30];
 
     unsigned char black_piece_pos[60];
     bool black_piece_alive[30];
     signed char black_piece_jump_moves[30][30];
     signed char black_piece_move_directions[30][28];
-    int black_piece_fist_move[30];
+    short black_piece_fist_move[30];
 
     unsigned char piece_count;
     bool boarder_x[30];
